When completing, insert space to fix callout
As it turns out, `> [!TYPE]TITLE` is not a valid callout.
The syntax is distinct enough that it's fair to assume that the
user *intended* for it to be a callout, though.

We can help the user out by inserting a space at the end of the
foldable part in the header, before the title.